(chat)[CN-60] update chat avatar
