+++ src/shared/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useBrandColors'
+export * from './useCloudinaryWidget'
 export * from './useDebounceValue'
 export * from './useErrorMessage'
