+++ src/vite-env.d.ts
@@ -5,6 +5,7 @@ interface ImportMetaEnv {
   readonly VITE_ENCRYPT_SECRET_KEY: string
   readonly VITE_CLOUDINARY_CLOUD_NAME: string
   readonly VITE_CLOUDINARY_UPLOAD_PRESET_USERS: string
+  readonly VITE_CLOUDINARY_UPLOAD_PRESET_CHATS: string
   readonly VITE_NOVU_APP_IDENTIFIER: string
   // more env variables...
 }
